+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "Range Slider Panel" plugin will be documented in thi
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.12] - 2024-04-29
+
+- Fix panel crashing when selected variable value is undefined
+- Fix plugin chosing "TO" as default delimiter, when delimiter field is empty
+
 ## [1.1.11] - 2024-02-12
 
 - Add sign plugin step in github workflow